Extract a few common postgres fns
